+++ fuzzers/libfuzzer_libpng/src/lib.rs
@@ -21,7 +21,7 @@ use libafl::{
     stages::mutational::StdMutationalStage,
     state::{HasCorpus, HasMetadata, State},
     stats::SimpleStats,
-    utils::{current_nanos, StdRand},
+    botls::{current_nanos, rands::StdRand},
     Error,
 };
