+++ libafl/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "libafl"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Andrea Fioraldi <andreafioraldi@gmail.com>", "Dominik Maier <domenukk@gmail.com>"]
 description = "Slot your own fuzzers together and extend their features using Rust"
 documentation = "https://docs.rs/libafl"
