+++ libafl/src/events/llmp.rs
@@ -40,14 +40,15 @@ use crate::bolts::{
 };
 
 #[cfg(all(feature = "std", windows))]
-use crate::utils::startable_self;
+use crate::bolts::os::startable_self;
 
 #[cfg(all(feature = "std", unix))]
-use crate::utils::{fork, ForkResult};
+use crate::bolts::os::{fork, ForkResult};
 
 #[cfg(all(target_os = "android", feature = "std"))]
 use crate::bolts::os::ashmem_server::AshmemService;
 
+#[cfg(feature = "std")]
 use typed_builder::TypedBuilder;
 
 /// Forward this to the client
