+++ libafl_frida/Cargo.toml
@@ -14,8 +14,8 @@ edition = "2018"
 cc = { version = "1.0", features = ["parallel"] }
 
 [dependencies]
-libafl = { path = "../libafl", version = "0.2.1", features = ["std", "libafl_derive"] }
-libafl_targets = { path = "../libafl_targets", version = "0.2.1" }
+libafl = { path = "../libafl", version = "0.3.0", features = ["std", "libafl_derive"] }
+libafl_targets = { path = "../libafl_targets", version = "0.3.0" }
 nix = "0.20.0"
 libc = "0.2.92"
 hashbrown = "0.11"
