+++ libafl_targets/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "libafl_targets"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Andrea Fioraldi <andreafioraldi@gmail.com>"]
 description = "Common code for target instrumentation that can be used combined with LibAFL"
 documentation = "https://docs.rs/libafl_targets"
